feat: convert from `torch` to `numpy/numba`
The visibility graph generation highly depends on `for`-loops, and
data-specific accesses with conditional statements. This makes it
unsuitable for GPU acceleration and mappings like `torch.vmap()`.

Therefore, to increase the performance, the code is converted from
`torch` to `numpy` with `numba` JIT compilation. This conversion highly
improves the performance of the code, even with the overhead of
GPU-to-CPU-to-GPU data transfer.