refactor: simplify code structure and rename
Significantly refactor the code structure to make it more modular and
shorter. Also set the line-length to 120 to eliminate redundant newlines
caused by argument splitting.

Export the weight calculation function to a separate file for future
alternative implementations.

Did not change any functionality, but also eliminate the theoretical
derivation notations from the code to make it more readable and idiomatic.