refactor(ts2vg): export naive `ts2vg` sanity check
The private functions of `ts2vg` are used to basic VVG's in a naive way.
These implementations are used for sanity checks of current VVG
implementations. To indicate that these functions are not intended for
users, they are exported to a file with prefixed with an underscore, and
only used in test files.